rename density_ve to ve_norm_gradh
