+++ sph/include/sph/hydro_ve/ve_norm_gradh_kern.hpp
